feat!: add --ignore to ignore MD5 checks
This is breaking as the --force option now enforce MD5 checks by default, which it didn't do previously.
